+++ contact/forms.py
@@ -5,10 +5,12 @@ from django.utils.translation import gettext_lazy as _
 from django.forms import ValidationError
 from requests import post
 
+from portal.forms import HealthcareBaseForm
+
 logger = logging.getLogger(__name__)
 
 
-class ContactForm(forms.Form):
+class ContactForm(HealthcareBaseForm):
     name = forms.CharField(label=_("Your name"))
     email = forms.EmailField(
         label=_("Your email address"),
@@ -29,20 +31,20 @@ class ContactForm(forms.Form):
                 settings.FRESHDESK_API_ENDPOINT + "/tickets",
                 json={
                     "email": self.cleaned_data.get("email", None),
-                    "subject": self.cleaned_data.get("feedback", "")[0:20],
+                    "subject": self.cleaned_data.get("feedback", "")[0:20] + "...",
                     "description": self.cleaned_data.get("feedback", ""),
                     "name": self.cleaned_data.get("name", ""),
                     "product_id": settings.FRESHDESK_PRODUCT_ID,
                     "status": 2,
                     "priority": 1,
                 },
-                auth=(settings.FRESHDESK_API_KEY, "dummyString"),
+                auth=(settings.FRESHDESK_API_KEY, ""),
             )
             response.raise_for_status()
         except Exception as e:
             logger.critical(f"Unable to communicate with Freshdesk : {e}")
             raise ValidationError(
                 _(
-                    "We are unable to send your feedback at the moment. Please try again."
+                    "There was a problem and your feedback wasn't submitted. Please email your feedback to assistance+covid@cds-snc.ca."
                 )
             )
