+++ portal/templates/includes/nav_main.html
@@ -7,6 +7,7 @@
       {% if user.is_admin %}
         <li><a {% if request.resolver_match.url_name == 'profiles' %}class="active"{% endif %} href="{% url 'profiles' %}">{% trans "Manage accounts" %}</a></li>
       {% endif %}
+      <li><a {% if request.resolver_match.url_name == 'index' %}class="active"{% endif %} href="{% url 'contact:index' %}">{% trans "Contact us" %}</a></li>
     </ul>
   </div>
 </nav>
