+++ profiles/forms.py
@@ -18,21 +18,9 @@ from invitations.forms import (
     InvitationAdminChangeForm,
 )
 
+from portal.forms import HealthcareBaseForm
 from .models import HealthcareUser, HealthcareProvince
 from .utils import generate_2fa_code
-
-
-class HealthcareBaseForm(forms.Form):
-    def __init__(self, *args, **kwargs):
-        # Remove the colon after field labels
-        kwargs.setdefault("label_suffix", "")
-        super().__init__(*args, **kwargs)
-
-        # override field attributes: https://stackoverflow.com/a/56870308
-        for field in self.fields:
-            self.fields[field].widget.attrs.pop("autofocus", None)
-            # remove autocomplete attributes
-            self.fields[field].widget.attrs.update({"autocomplete": "off"})
 
 
 class HealthcareAuthenticationForm(HealthcareBaseForm, AuthenticationForm):
